signup and signout completed

+++ src/actions/auth.actions.js
@@ -1,69 +1,79 @@
-import {authConstants} from './constants';
+import { authConstants } from './constants';
 import axios from '../helpers/axios';
 
 export const login = (user) => {
-
   // console.log(user);
 
   return async (dispatch) => {
-    dispatch({type: authConstants.LOGIN_REQUEST});
+    dispatch({ type: authConstants.LOGIN_REQUEST });
     const res = await axios.post('/admin/signin', {
-      ...user
+      ...user,
     });
 
-    if(res.status === 200){
-      const {token, user} = res.data;
+    if (res.status === 200) {
+      const { token, user } = res.data;
       localStorage.setItem('token', token);
       localStorage.setItem('user', JSON.stringify(user));
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
-          token, user
-        }
+          token,
+          user,
+        },
       });
-    }else{
-      if(res.status === 400){
+    } else {
+      if (res.status === 400) {
         dispatch({
-          type: authConstants.LOGIN_SUCCESS,
+          type: authConstants.LOGIN_FAILURE,
           payload: {
-            error: res.data.error
-          }
+            error: res.data.error,
+          },
         });
       }
-
     }
-
-  }
-}
+  };
+};
 
 export const isUserLoggedIn = () => {
-  return async dispatch => {
+  return async (dispatch) => {
     const token = localStorage.getItem('token');
-    if(token){
+    if (token) {
       const user = JSON.parse(localStorage.getItem('user'));
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
-          token, user
-        }
+          token,
+          user,
+        },
       });
-    }else{
+    } else {
       dispatch({
         type: authConstants.LOGIN_FAILURE,
         payload: {
-          error: 'Failed to Login'
-        }
+          error: 'Failed to Login',
+        },
       });
     }
-  }
-}
+  };
+};
 
-export const signout = ()=>{
-  return async dispatch => {
-    console.log("Inside Signout");
-    localStorage.clear();
+export const signout = () => {
+  return async (dispatch) => {
     dispatch({
-      type: authConstants.LOGOUT_REQUEST
-    })
-  }
-}
+      type: authConstants.LOGOUT_REQUEST,
+    });
+    const res = await axios.post('/admin/signout');
+    console.log(res);
+    if (res.status === 200 || 500) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      });
+    }else{
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: {error: res.data.error}
+      });
+    }
+  };
+};
